+++ inc/Engine/Activation/Activation.php
@@ -4,8 +4,6 @@ namespace WP_Rocket\Engine\Activation;
 
 use WP_Rocket\Admin\Options;
 use WP_Rocket\Dependencies\League\Container\Container;
-use WP_Rocket\Engine\Preload\ActivationServiceProvider;
-use WP_Rocket\Engine\Preload\Subscriber;
 use WP_Rocket\Event_Management\Event_Manager;
 use WP_Rocket\ThirdParty\Hostings\HostResolver;
