+++ inc/Engine/Preload/Admin/Settings.php
@@ -80,6 +80,15 @@ class Settings {
 	 * @return void
 	 */
 	public function maybe_display_as_missed_tables_notice() {
+
+		if ( function_exists( 'get_current_screen' ) && 'tools_page_action-scheduler' === get_current_screen()->id ) {
+			return;
+		}
+
+		if ( $this->is_valid_as_tables() ) {
+			return;
+		}
+
 		$as_tools_link = menu_page_url( 'action-scheduler', false );
 		$message       = sprintf(
 		// translators: %1$s = plugin name, %2$s = opening anchor tag, %3$s = closing anchor tag.
@@ -104,5 +113,30 @@ class Settings {
 	 */
 	public function is_enabled() : bool {
 		return (bool) $this->options->get( 'manual_preload', 0 );
+	}
+
+	/**
+	 * Checks if Action scheduler tables are there or not.
+	 *
+	 * @since 3.11.0.3
+	 *
+	 * @return bool
+	 */
+	private function is_valid_as_tables() {
+		$cached_count = get_transient( 'rocket_preload_as_tables_count' );
+		if ( false !== $cached_count && ! is_admin() ) { // Stop caching in admin UI.
+			return 4 === (int) $cached_count;
+		}
+
+		global $wpdb;
+
+		// phpcs:ignore WordPress.DB.DirectDatabaseQuery.DirectQuery,WordPress.DB.DirectDatabaseQuery.NoCaching
+		$found_as_tables = $wpdb->get_col(
+			$wpdb->prepare( 'SHOW TABLES LIKE %s', $wpdb->prefix . 'actionscheduler%' )
+		);
+
+		set_transient( 'rocket_preload_as_tables_count', count( $found_as_tables ), rocket_get_constant( 'DAY_IN_SECONDS', 24 * 60 * 60 ) );
+
+		return 4 === count( $found_as_tables );
 	}
 }
