+++ inc/Engine/Preload/Database/Queries/Cache.php
@@ -425,6 +425,12 @@ class Cache extends Query {
 
 		$prefixed_table_name = $db->prefix . $this->table_name;
 
+		/**
+		 * Filter condition for cleaning URLS in the database.
+		 *
+		 * @param string $condition condition for cleaning URLS in the database.
+		 * @returns string
+		 */
 		$condition = apply_filters( 'rocket_preload_all_to_pending_condition', ' WHERE 1 = 1' );
 
 		$db->query( "UPDATE `$prefixed_table_name` SET status = 'pending'$condition" );
@@ -448,7 +454,7 @@ class Cache extends Query {
 	}
 
 	/**
-	 * Check if the page is preloaded.
+	 * Check if the page is pending.
 	 *
 	 * @param string $url url from the page to check.
 	 * @return bool
